refactor(core): move file system endpoints to its own plugin
