+++ .changes/move-fs.md
@@ -1,0 +1,6 @@
+---
+"api": patch
+"tauri": patch
+---
+
+Moved the file system APIs to its own plugin in the plugins-workspace repository.
