+++ core/tauri/scripts/bundle.global.js
@@ -1,7 +1,7 @@
-"use strict";var __TAURI_IIFE__=(()=>{var L=Object.defineProperty;var le=Object.getOwnPropertyDescriptor;var ue=Object.getOwnPropertyNames;var de=Object.prototype.hasOwnProperty;var c=(n,e)=>{for(var t in e)L(n,t,{get:e[t],enumerable:!0})},ce=(n,e,t,r)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of ue(e))!de.call(n,s)&&s!==t&&L(n,s,{get:()=>e[s],enumerable:!(r=le(e,s))||r.enumerable});return n};var me=n=>ce(L({},"__esModule",{value:!0}),n);var jt={};c(jt,{app:()=>R,clipboard:()=>U,dialog:()=>N,event:()=>H,fs:()=>V,globalShortcut:()=>j,http:()=>G,invoke:()=>Vt,notification:()=>$,os:()=>te,path:()=>J,process:()=>K,shell:()=>Q,tauri:()=>k,updater:()=>Z,window:()=>ee});var R={};c(R,{getName:()=>he,getTauriVersion:()=>fe,getVersion:()=>ye,hide:()=>_e,show:()=>be});var k={};c(k,{convertFileSrc:()=>ge,invoke:()=>l,transformCallback:()=>m});function pe(){return window.crypto.getRandomValues(new Uint32Array(1))[0]}function m(n,e=!1){let t=pe(),r=`_${t}`;return Object.defineProperty(window,r,{value:s=>(e&&Reflect.deleteProperty(window,r),n?.(s)),writable:!1,configurable:!0}),t}async function l(n,e={}){return new Promise((t,r)=>{let s=m(u=>{t(u),Reflect.deleteProperty(window,`_${o}`)},!0),o=m(u=>{r(u),Reflect.deleteProperty(window,`_${s}`)},!0);window.__TAURI_IPC__({cmd:n,callback:s,error:o,...e})})}function ge(n,e="asset"){let t=encodeURIComponent(n);return navigator.userAgent.includes("Windows")?`https://${e}.localhost/${t}`:`${e}://localhost/${t}`}async function i(n){return l("tauri",n)}async function ye(){return i({__tauriModule:"App",message:{cmd:"getAppVersion"}})}async function he(){return i({__tauriModule:"App",message:{cmd:"getAppName"}})}async function fe(){return i({__tauriModule:"App",message:{cmd:"getTauriVersion"}})}async function be(){return i({__tauriModule:"App",message:{cmd:"show"}})}async function _e(){return i({__tauriModule:"App",message:{cmd:"hide"}})}var U={};c(U,{readText:()=>we,writeText:()=>Pe});async function Pe(n){return i({__tauriModule:"Clipboard",message:{cmd:"writeText",data:n}})}async function we(){return i({__tauriModule:"Clipboard",message:{cmd:"readText",data:null}})}var N={};c(N,{ask:()=>Ee,confirm:()=>Fe,message:()=>Te,open:()=>ve,save:()=>Oe});async function ve(n={}){return typeof n=="object"&&Object.freeze(n),i({__tauriModule:"Dialog",message:{cmd:"openDialog",options:n}})}async function Oe(n={}){return typeof n=="object"&&Object.freeze(n),i({__tauriModule:"Dialog",message:{cmd:"saveDialog",options:n}})}async function Te(n,e){let t=typeof e=="string"?{title:e}:e;return i({__tauriModule:"Dialog",message:{cmd:"messageDialog",message:n.toString(),title:t?.title?.toString(),type:t?.type,buttonLabel:t?.okLabel?.toString()}})}async function Ee(n,e){let t=typeof e=="string"?{title:e}:e;return i({__tauriModule:"Dialog",message:{cmd:"askDialog",message:n.toString(),title:t?.title?.toString(),type:t?.type,buttonLabels:[t?.okLabel?.toString()??"Yes",t?.cancelLabel?.toString()??"No"]}})}async function Fe(n,e){let t=typeof e=="string"?{title:e}:e;return i({__tauriModule:"Dialog",message:{cmd:"confirmDialog",message:n.toString(),title:t?.title?.toString(),type:t?.type,buttonLabels:[t?.okLabel?.toString()??"Ok",t?.cancelLabel?.toString()??"Cancel"]}})}var H={};c(H,{TauriEvent:()=>O,emit:()=>T,listen:()=>I,once:()=>z});async function ne(n,e){return i({__tauriModule:"Event",message:{cmd:"unlisten",event:n,eventId:e}})}async function w(n,e,t){await i({__tauriModule:"Event",message:{cmd:"emit",event:n,windowLabel:e,payload:t}})}async function _(n,e,t){return i({__tauriModule:"Event",message:{cmd:"listen",event:n,windowLabel:e,handler:m(t)}}).then(r=>async()=>ne(n,r))}async function v(n,e,t){return _(n,e,r=>{t(r),ne(n,r.id).catch(()=>{})})}var O=(d=>(d.WINDOW_RESIZED="tauri://resize",d.WINDOW_MOVED="tauri://move",d.WINDOW_CLOSE_REQUESTED="tauri://close-requested",d.WINDOW_CREATED="tauri://window-created",d.WINDOW_DESTROYED="tauri://destroyed",d.WINDOW_FOCUS="tauri://focus",d.WINDOW_BLUR="tauri://blur",d.WINDOW_SCALE_FACTOR_CHANGED="tauri://scale-change",d.WINDOW_THEME_CHANGED="tauri://theme-changed",d.WINDOW_FILE_DROP="tauri://file-drop",d.WINDOW_FILE_DROP_HOVER="tauri://file-drop-hover",d.WINDOW_FILE_DROP_CANCELLED="tauri://file-drop-cancelled",d.MENU="tauri://menu",d.CHECK_UPDATE="tauri://update",d.UPDATE_AVAILABLE="tauri://update-available",d.INSTALL_UPDATE="tauri://update-install",d.STATUS_UPDATE="tauri://update-status",d.DOWNLOAD_PROGRESS="tauri://update-download-progress",d))(O||{});async function I(n,e){return _(n,null,e)}async function z(n,e){return v(n,null,e)}async function T(n,e){return w(n,void 0,e)}var V={};c(V,{BaseDirectory:()=>E,Dir:()=>E,copyFile:()=>ke,createDir:()=>We,exists:()=>Ne,readBinaryFile:()=>Ae,readDir:()=>Se,readTextFile:()=>Ce,removeDir:()=>Le,removeFile:()=>Re,renameFile:()=>Ue,writeBinaryFile:()=>xe,writeFile:()=>De,writeTextFile:()=>De});var E=(a=>(a[a.Audio=1]="Audio",a[a.Cache=2]="Cache",a[a.Config=3]="Config",a[a.Data=4]="Data",a[a.LocalData=5]="LocalData",a[a.Document=6]="Document",a[a.Download=7]="Download",a[a.Picture=8]="Picture",a[a.Public=9]="Public",a[a.Video=10]="Video",a[a.Resource=11]="Resource",a[a.Temp=12]="Temp",a[a.AppConfig=13]="AppConfig",a[a.AppData=14]="AppData",a[a.AppLocalData=15]="AppLocalData",a[a.AppCache=16]="AppCache",a[a.AppLog=17]="AppLog",a[a.Desktop=18]="Desktop",a[a.Executable=19]="Executable",a[a.Font=20]="Font",a[a.Home=21]="Home",a[a.Runtime=22]="Runtime",a[a.Template=23]="Template",a))(E||{});async function Ce(n,e={}){return i({__tauriModule:"Fs",message:{cmd:"readTextFile",path:n,options:e}})}async function Ae(n,e={}){let t=await i({__tauriModule:"Fs",message:{cmd:"readFile",path:n,options:e}});return Uint8Array.from(t)}async function De(n,e,t){typeof t=="object"&&Object.freeze(t),typeof n=="object"&&Object.freeze(n);let r={path:"",contents:""},s=t;return typeof n=="string"?r.path=n:(r.path=n.path,r.contents=n.contents),typeof e=="string"?r.contents=e??"":s=e,i({__tauriModule:"Fs",message:{cmd:"writeFile",path:r.path,contents:Array.from(new TextEncoder().encode(r.contents)),options:s}})}async function xe(n,e,t){typeof t=="object"&&Object.freeze(t),typeof n=="object"&&Object.freeze(n);let r={path:"",contents:[]},s=t;return typeof n=="string"?r.path=n:(r.path=n.path,r.contents=n.contents),e&&"dir"in e?s=e:typeof n=="string"&&(r.contents=e??[]),i({__tauriModule:"Fs",message:{cmd:"writeFile",path:r.path,contents:Array.from(r.contents instanceof ArrayBuffer?new Uint8Array(r.contents):r.contents),options:s}})}async function Se(n,e={}){return i({__tauriModule:"Fs",message:{cmd:"readDir",path:n,options:e}})}async function We(n,e={}){return i({__tauriModule:"Fs",message:{cmd:"createDir",path:n,options:e}})}async function Le(n,e={}){return i({__tauriModule:"Fs",message:{cmd:"removeDir",path:n,options:e}})}async function ke(n,e,t={}){return i({__tauriModule:"Fs",message:{cmd:"copyFile",source:n,destination:e,options:t}})}async function Re(n,e={}){return i({__tauriModule:"Fs",message:{cmd:"removeFile",path:n,options:e}})}async function Ue(n,e,t={}){return i({__tauriModule:"Fs",message:{cmd:"renameFile",oldPath:n,newPath:e,options:t}})}async function Ne(n,e={}){return i({__tauriModule:"Fs",message:{cmd:"exists",path:n,options:e}})}var j={};c(j,{isRegistered:()=>He,register:()=>Ie,registerAll:()=>ze,unregister:()=>Ve,unregisterAll:()=>je});async function Ie(n,e){return i({__tauriModule:"GlobalShortcut",message:{cmd:"register",shortcut:n,handler:m(e)}})}async function ze(n,e){return i({__tauriModule:"GlobalShortcut",message:{cmd:"registerAll",shortcuts:n,handler:m(e)}})}async function He(n){return i({__tauriModule:"GlobalShortcut",message:{cmd:"isRegistered",shortcut:n}})}async function Ve(n){return i({__tauriModule:"GlobalShortcut",message:{cmd:"unregister",shortcut:n}})}async function je(){return i({__tauriModule:"GlobalShortcut",message:{cmd:"unregisterAll"}})}var G={};c(G,{Body:()=>g,Client:()=>M,Response:()=>F,ResponseType:()=>ie,fetch:()=>qe,getClient:()=>re});var ie=(r=>(r[r.JSON=1]="JSON",r[r.Text=2]="Text",r[r.Binary=3]="Binary",r))(ie||{}),g=class{constructor(e,t){this.type=e,this.payload=t}static form(e){let t={},r=(s,o)=>{if(o!==null){let u;typeof o=="string"?u=o:o instanceof Uint8Array||Array.isArray(o)?u=Array.from(o):o instanceof File?u={file:o.name,mime:o.type,fileName:o.name}:typeof o.file=="string"?u={file:o.file,mime:o.mime,fileName:o.fileName}:u={file:Array.from(o.file),mime:o.mime,fileName:o.fileName},t[String(s)]=u}};if(e instanceof FormData)for(let[s,o]of e)r(s,o);else for(let[s,o]of Object.entries(e))r(s,o);return new g("Form",t)}static json(e){return new g("Json",e)}static text(e){return new g("Text",e)}static bytes(e){return new g("Bytes",Array.from(e instanceof ArrayBuffer?new Uint8Array(e):e))}},F=class{constructor(e){this.url=e.url,this.status=e.status,this.ok=this.status>=200&&this.status<300,this.headers=e.headers,this.rawHeaders=e.rawHeaders,this.data=e.data}},M=class{constructor(e){this.id=e}async drop(){return i({__tauriModule:"Http",message:{cmd:"dropClient",client:this.id}})}async request(e){let t=!e.responseType||e.responseType===1;return t&&(e.responseType=2),i({__tauriModule:"Http",message:{cmd:"httpRequest",client:this.id,options:e}}).then(r=>{let s=new F(r);if(t){try{s.data=JSON.parse(s.data)}catch(o){if(s.ok&&s.data==="")s.data={};else if(s.ok)throw Error(`Failed to parse response \`${s.data}\` as JSON: ${o};
-              try setting the \`responseType\` option to \`ResponseType.Text\` or \`ResponseType.Binary\` if the API does not return a JSON response.`)}return s}return s})}async get(e,t){return this.request({method:"GET",url:e,...t})}async post(e,t,r){return this.request({method:"POST",url:e,body:t,...r})}async put(e,t,r){return this.request({method:"PUT",url:e,body:t,...r})}async patch(e,t){return this.request({method:"PATCH",url:e,...t})}async delete(e,t){return this.request({method:"DELETE",url:e,...t})}};async function re(n){return i({__tauriModule:"Http",message:{cmd:"createClient",options:n}}).then(e=>new M(e))}var q=null;async function qe(n,e){return q===null&&(q=await re()),q.request({url:n,method:e?.method??"GET",...e})}var $={};c($,{isPermissionGranted:()=>Ge,requestPermission:()=>$e,sendNotification:()=>Je});async function Ge(){return window.Notification.permission!=="default"?Promise.resolve(window.Notification.permission==="granted"):i({__tauriModule:"Notification",message:{cmd:"isNotificationPermissionGranted"}})}async function $e(){return window.Notification.requestPermission()}function Je(n){typeof n=="string"?new window.Notification(n):new window.Notification(n.title,n)}var J={};c(J,{BaseDirectory:()=>E,appCacheDir:()=>Ze,appConfigDir:()=>Ke,appDataDir:()=>Qe,appLocalDataDir:()=>Ye,appLogDir:()=>ht,audioDir:()=>Xe,basename:()=>Tt,cacheDir:()=>Be,configDir:()=>et,dataDir:()=>tt,delimiter:()=>bt,desktopDir:()=>nt,dirname:()=>vt,documentDir:()=>it,downloadDir:()=>rt,executableDir:()=>st,extname:()=>Ot,fontDir:()=>ot,homeDir:()=>at,isAbsolute:()=>Et,join:()=>wt,localDataDir:()=>lt,normalize:()=>Pt,pictureDir:()=>ut,publicDir:()=>dt,resolve:()=>_t,resolveResource:()=>mt,resourceDir:()=>ct,runtimeDir:()=>pt,sep:()=>ft,templateDir:()=>gt,videoDir:()=>yt});function P(){return navigator.appVersion.includes("Win")}async function Ke(){return l("plugin:path|resolve_directory",{directory:13})}async function Qe(){return l("plugin:path|resolve_directory",{directory:14})}async function Ye(){return l("plugin:path|resolve_directory",{directory:15})}async function Ze(){return l("plugin:path|resolve_directory",{directory:16})}async function Xe(){return l("plugin:path|resolve_directory",{directory:1})}async function Be(){return l("plugin:path|resolve_directory",{directory:2})}async function et(){return l("plugin:path|resolve_directory",{directory:3})}async function tt(){return l("plugin:path|resolve_directory",{directory:4})}async function nt(){return l("plugin:path|resolve_directory",{directory:18})}async function it(){return l("plugin:path|resolve_directory",{directory:6})}async function rt(){return l("plugin:path|resolve_directory",{directory:7})}async function st(){return l("plugin:path|resolve_directory",{directory:19})}async function ot(){return l("plugin:path|resolve_directory",{directory:20})}async function at(){return l("plugin:path|resolve_directory",{directory:21})}async function lt(){return l("plugin:path|resolve_directory",{directory:5})}async function ut(){return l("plugin:path|resolve_directory",{directory:8})}async function dt(){return l("plugin:path|resolve_directory",{directory:9})}async function ct(){return l("plugin:path|resolve_directory",{directory:11})}async function mt(n){return l("plugin:path|resolve_directory",{directory:11,path:n})}async function pt(){return l("plugin:path|resolve_directory",{directory:22})}async function gt(){return l("plugin:path|resolve_directory",{directory:23})}async function yt(){return l("plugin:path|resolve_directory",{directory:10})}async function ht(){return l("plugin:path|resolve_directory",{directory:17})}var ft=P()?"\\":"/",bt=P()?";":":";async function _t(...n){return l("plugin:path|resolve",{paths:n})}async function Pt(n){return l("plugin:path|normalize",{path:n})}async function wt(...n){return l("plugin:path|join",{paths:n})}async function vt(n){return l("plugin:path|dirname",{path:n})}async function Ot(n){return l("plugin:path|extname",{path:n})}async function Tt(n,e){return l("plugin:path|basename",{path:n,ext:e})}async function Et(n){return l("plugin:path|isAbsolute",{path:n})}var K={};c(K,{exit:()=>Ft,relaunch:()=>Mt});async function Ft(n=0){return i({__tauriModule:"Process",message:{cmd:"exit",exitCode:n}})}async function Mt(){return i({__tauriModule:"Process",message:{cmd:"relaunch"}})}var Q={};c(Q,{Child:()=>C,Command:()=>h,EventEmitter:()=>y,open:()=>At});async function Ct(n,e,t=[],r){return typeof t=="object"&&Object.freeze(t),i({__tauriModule:"Shell",message:{cmd:"execute",program:e,args:t,options:r,onEventFn:m(n)}})}var y=class{constructor(){this.eventListeners=Object.create(null)}addListener(e,t){return this.on(e,t)}removeListener(e,t){return this.off(e,t)}on(e,t){return e in this.eventListeners?this.eventListeners[e].push(t):this.eventListeners[e]=[t],this}once(e,t){let r=s=>{this.removeListener(e,r),t(s)};return this.addListener(e,r)}off(e,t){return e in this.eventListeners&&(this.eventListeners[e]=this.eventListeners[e].filter(r=>r!==t)),this}removeAllListeners(e){return e?delete this.eventListeners[e]:this.eventListeners=Object.create(null),this}emit(e,t){if(e in this.eventListeners){let r=this.eventListeners[e];for(let s of r)s(t);return!0}return!1}listenerCount(e){return e in this.eventListeners?this.eventListeners[e].length:0}prependListener(e,t){return e in this.eventListeners?this.eventListeners[e].unshift(t):this.eventListeners[e]=[t],this}prependOnceListener(e,t){let r=s=>{this.removeListener(e,r),t(s)};return this.prependListener(e,r)}},C=class{constructor(e){this.pid=e}async write(e){return i({__tauriModule:"Shell",message:{cmd:"stdinWrite",pid:this.pid,buffer:typeof e=="string"?e:Array.from(e)}})}async kill(){return i({__tauriModule:"Shell",message:{cmd:"killChild",pid:this.pid}})}},h=class extends y{constructor(t,r=[],s){super();this.stdout=new y;this.stderr=new y;this.program=t,this.args=typeof r=="string"?[r]:r,this.options=s??{}}static create(t,r=[],s){return new h(t,r,s)}static sidecar(t,r=[],s){let o=new h(t,r,s);return o.options.sidecar=!0,o}async spawn(){return Ct(t=>{switch(t.event){case"Error":this.emit("error",t.payload);break;case"Terminated":this.emit("close",t.payload);break;case"Stdout":this.stdout.emit("data",t.payload);break;case"Stderr":this.stderr.emit("data",t.payload);break}},this.program,this.args,this.options).then(t=>new C(t))}async execute(){return new Promise((t,r)=>{this.on("error",r);let s=[],o=[];this.stdout.on("data",u=>{s.push(u)}),this.stderr.on("data",u=>{o.push(u)}),this.on("close",u=>{t({code:u.code,signal:u.signal,stdout:this.collectOutput(s),stderr:this.collectOutput(o)})}),this.spawn().catch(r)})}collectOutput(t){return this.options.encoding==="raw"?t.reduce((r,s)=>new Uint8Array([...r,...s,10]),new Uint8Array):t.join(`
-`)}};async function At(n,e){return i({__tauriModule:"Shell",message:{cmd:"open",path:n,with:e}})}var Z={};c(Z,{checkUpdate:()=>xt,installUpdate:()=>Dt,onUpdaterEvent:()=>Y});async function Y(n){return I("tauri://update-status",e=>{n(e?.payload)})}async function Dt(){let n;function e(){n&&n(),n=void 0}return new Promise((t,r)=>{function s(o){if(o.error){e(),r(o.error);return}o.status==="DONE"&&(e(),t())}Y(s).then(o=>{n=o}).catch(o=>{throw e(),o}),T("tauri://update-install").catch(o=>{throw e(),o})})}async function xt(){let n;function e(){n&&n(),n=void 0}return new Promise((t,r)=>{function s(u){e(),t({manifest:u,shouldUpdate:!0})}function o(u){if(u.error){e(),r(u.error);return}u.status==="UPTODATE"&&(e(),t({shouldUpdate:!1}))}z("tauri://update-available",u=>{s(u?.payload)}).catch(u=>{throw e(),u}),Y(o).then(u=>{n=u}).catch(u=>{throw e(),u}),T("tauri://update").catch(u=>{throw e(),u})})}var ee={};c(ee,{CloseRequestedEvent:()=>W,LogicalPosition:()=>D,LogicalSize:()=>A,PhysicalPosition:()=>b,PhysicalSize:()=>f,UserAttentionType:()=>oe,WebviewWindow:()=>p,WebviewWindowHandle:()=>x,WindowManager:()=>S,appWindow:()=>X,availableMonitors:()=>kt,currentMonitor:()=>Wt,getAll:()=>ae,getCurrent:()=>St,primaryMonitor:()=>Lt});var A=class{constructor(e,t){this.type="Logical";this.width=e,this.height=t}},f=class{constructor(e,t){this.type="Physical";this.width=e,this.height=t}toLogical(e){return new A(this.width/e,this.height/e)}},D=class{constructor(e,t){this.type="Logical";this.x=e,this.y=t}},b=class{constructor(e,t){this.type="Physical";this.x=e,this.y=t}toLogical(e){return new D(this.x/e,this.y/e)}},oe=(t=>(t[t.Critical=1]="Critical",t[t.Informational=2]="Informational",t))(oe||{});function St(){return new p(window.__TAURI_METADATA__.__currentWindow.label,{skip:!0})}function ae(){return window.__TAURI_METADATA__.__windows.map(n=>new p(n.label,{skip:!0}))}var se=["tauri://created","tauri://error"],x=class{constructor(e){this.label=e,this.listeners=Object.create(null)}async listen(e,t){return this._handleTauriEvent(e,t)?Promise.resolve(()=>{let r=this.listeners[e];r.splice(r.indexOf(t),1)}):_(e,this.label,t)}async once(e,t){return this._handleTauriEvent(e,t)?Promise.resolve(()=>{let r=this.listeners[e];r.splice(r.indexOf(t),1)}):v(e,this.label,t)}async emit(e,t){if(se.includes(e)){for(let r of this.listeners[e]||[])r({event:e,id:-1,windowLabel:this.label,payload:t});return Promise.resolve()}return w(e,this.label,t)}_handleTauriEvent(e,t){return se.includes(e)?(e in this.listeners?this.listeners[e].push(t):this.listeners[e]=[t],!0):!1}},S=class extends x{async scaleFactor(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"scaleFactor"}}}})}async innerPosition(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"innerPosition"}}}}).then(({x:e,y:t})=>new b(e,t))}async outerPosition(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"outerPosition"}}}}).then(({x:e,y:t})=>new b(e,t))}async innerSize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"innerSize"}}}}).then(({width:e,height:t})=>new f(e,t))}async outerSize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"outerSize"}}}}).then(({width:e,height:t})=>new f(e,t))}async isFullscreen(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isFullscreen"}}}})}async isMinimized(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isMinimized"}}}})}async isMaximized(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isMaximized"}}}})}async isDecorated(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isDecorated"}}}})}async isResizable(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isResizable"}}}})}async isVisible(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isVisible"}}}})}async title(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"title"}}}})}async theme(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"theme"}}}})}async center(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"center"}}}})}async requestUserAttention(e){let t=null;return e&&(e===1?t={type:"Critical"}:t={type:"Informational"}),i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"requestUserAttention",payload:t}}}})}async setResizable(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setResizable",payload:e}}}})}async setTitle(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setTitle",payload:e}}}})}async maximize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"maximize"}}}})}async unmaximize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"unmaximize"}}}})}async toggleMaximize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"toggleMaximize"}}}})}async minimize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"minimize"}}}})}async unminimize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"unminimize"}}}})}async show(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"show"}}}})}async hide(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"hide"}}}})}async close(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"close"}}}})}async setDecorations(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setDecorations",payload:e}}}})}async setShadow(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setShadow",payload:e}}}})}async setAlwaysOnTop(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setAlwaysOnTop",payload:e}}}})}async setContentProtected(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setContentProtected",payload:e}}}})}async setSize(e){if(!e||e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `size` argument must be either a LogicalSize or a PhysicalSize instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setSize",payload:{type:e.type,data:{width:e.width,height:e.height}}}}}})}async setMinSize(e){if(e&&e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `size` argument must be either a LogicalSize or a PhysicalSize instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setMinSize",payload:e?{type:e.type,data:{width:e.width,height:e.height}}:null}}}})}async setMaxSize(e){if(e&&e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `size` argument must be either a LogicalSize or a PhysicalSize instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setMaxSize",payload:e?{type:e.type,data:{width:e.width,height:e.height}}:null}}}})}async setPosition(e){if(!e||e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `position` argument must be either a LogicalPosition or a PhysicalPosition instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setPosition",payload:{type:e.type,data:{x:e.x,y:e.y}}}}}})}async setFullscreen(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setFullscreen",payload:e}}}})}async setFocus(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setFocus"}}}})}async setIcon(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setIcon",payload:{icon:typeof e=="string"?e:Array.from(e)}}}}})}async setSkipTaskbar(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setSkipTaskbar",payload:e}}}})}async setCursorGrab(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorGrab",payload:e}}}})}async setCursorVisible(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorVisible",payload:e}}}})}async setCursorIcon(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorIcon",payload:e}}}})}async setCursorPosition(e){if(!e||e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `position` argument must be either a LogicalPosition or a PhysicalPosition instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorPosition",payload:{type:e.type,data:{x:e.x,y:e.y}}}}}})}async setIgnoreCursorEvents(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setIgnoreCursorEvents",payload:e}}}})}async startDragging(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"startDragging"}}}})}async onResized(e){return this.listen("tauri://resize",e)}async onMoved(e){return this.listen("tauri://move",e)}async onCloseRequested(e){return this.listen("tauri://close-requested",t=>{let r=new W(t);Promise.resolve(e(r)).then(()=>{if(!r.isPreventDefault())return this.close()})})}async onFocusChanged(e){let t=await this.listen("tauri://focus",s=>{e({...s,payload:!0})}),r=await this.listen("tauri://blur",s=>{e({...s,payload:!1})});return()=>{t(),r()}}async onScaleChanged(e){return this.listen("tauri://scale-change",e)}async onMenuClicked(e){return this.listen("tauri://menu",e)}async onFileDropEvent(e){let t=await this.listen("tauri://file-drop",o=>{e({...o,payload:{type:"drop",paths:o.payload}})}),r=await this.listen("tauri://file-drop-hover",o=>{e({...o,payload:{type:"hover",paths:o.payload}})}),s=await this.listen("tauri://file-drop-cancelled",o=>{e({...o,payload:{type:"cancel"}})});return()=>{t(),r(),s()}}async onThemeChanged(e){return this.listen("tauri://theme-changed",e)}},W=class{constructor(e){this._preventDefault=!1;this.event=e.event,this.windowLabel=e.windowLabel,this.id=e.id}preventDefault(){this._preventDefault=!0}isPreventDefault(){return this._preventDefault}},p=class extends S{constructor(e,t={}){super(e),t?.skip||i({__tauriModule:"Window",message:{cmd:"createWebview",data:{options:{label:e,...t}}}}).then(async()=>this.emit("tauri://created")).catch(async r=>this.emit("tauri://error",r))}static getByLabel(e){return ae().some(t=>t.label===e)?new p(e,{skip:!0}):null}},X;"__TAURI_METADATA__"in window?X=new p(window.__TAURI_METADATA__.__currentWindow.label,{skip:!0}):(console.warn(`Could not find "window.__TAURI_METADATA__". The "appWindow" value will reference the "main" window label.
-Note that this is not an issue if running this frontend on a browser instead of a Tauri window.`),X=new p("main",{skip:!0}));function B(n){return n===null?null:{name:n.name,scaleFactor:n.scaleFactor,position:new b(n.position.x,n.position.y),size:new f(n.size.width,n.size.height)}}async function Wt(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{cmd:{type:"currentMonitor"}}}}).then(B)}async function Lt(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{cmd:{type:"primaryMonitor"}}}}).then(B)}async function kt(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{cmd:{type:"availableMonitors"}}}}).then(n=>n.map(B))}var te={};c(te,{EOL:()=>Rt,arch:()=>zt,platform:()=>Ut,tempdir:()=>Ht,type:()=>It,version:()=>Nt});var Rt=P()?`\r
+"use strict";var __TAURI_IIFE__=(()=>{var L=Object.defineProperty;var oe=Object.getOwnPropertyDescriptor;var le=Object.getOwnPropertyNames;var ue=Object.prototype.hasOwnProperty;var c=(n,e)=>{for(var t in e)L(n,t,{get:e[t],enumerable:!0})},de=(n,e,t,r)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of le(e))!ue.call(n,s)&&s!==t&&L(n,s,{get:()=>e[s],enumerable:!(r=oe(e,s))||r.enumerable});return n};var ce=n=>de(L({},"__esModule",{value:!0}),n);var Wt={};c(Wt,{app:()=>R,clipboard:()=>N,dialog:()=>U,event:()=>z,globalShortcut:()=>H,http:()=>q,invoke:()=>St,notification:()=>G,os:()=>B,path:()=>j,process:()=>$,shell:()=>J,tauri:()=>k,updater:()=>Q,window:()=>X});var R={};c(R,{getName:()=>ye,getTauriVersion:()=>he,getVersion:()=>ge,hide:()=>be,show:()=>fe});var k={};c(k,{convertFileSrc:()=>pe,invoke:()=>l,transformCallback:()=>m});function me(){return window.crypto.getRandomValues(new Uint32Array(1))[0]}function m(n,e=!1){let t=me(),r=`_${t}`;return Object.defineProperty(window,r,{value:s=>(e&&Reflect.deleteProperty(window,r),n?.(s)),writable:!1,configurable:!0}),t}async function l(n,e={}){return new Promise((t,r)=>{let s=m(u=>{t(u),Reflect.deleteProperty(window,`_${a}`)},!0),a=m(u=>{r(u),Reflect.deleteProperty(window,`_${s}`)},!0);window.__TAURI_IPC__({cmd:n,callback:s,error:a,...e})})}function pe(n,e="asset"){let t=encodeURIComponent(n);return navigator.userAgent.includes("Windows")?`https://${e}.localhost/${t}`:`${e}://localhost/${t}`}async function i(n){return l("tauri",n)}async function ge(){return i({__tauriModule:"App",message:{cmd:"getAppVersion"}})}async function ye(){return i({__tauriModule:"App",message:{cmd:"getAppName"}})}async function he(){return i({__tauriModule:"App",message:{cmd:"getTauriVersion"}})}async function fe(){return i({__tauriModule:"App",message:{cmd:"show"}})}async function be(){return i({__tauriModule:"App",message:{cmd:"hide"}})}var N={};c(N,{readText:()=>Pe,writeText:()=>_e});async function _e(n){return i({__tauriModule:"Clipboard",message:{cmd:"writeText",data:n}})}async function Pe(){return i({__tauriModule:"Clipboard",message:{cmd:"readText",data:null}})}var U={};c(U,{ask:()=>Te,confirm:()=>Ee,message:()=>Oe,open:()=>we,save:()=>ve});async function we(n={}){return typeof n=="object"&&Object.freeze(n),i({__tauriModule:"Dialog",message:{cmd:"openDialog",options:n}})}async function ve(n={}){return typeof n=="object"&&Object.freeze(n),i({__tauriModule:"Dialog",message:{cmd:"saveDialog",options:n}})}async function Oe(n,e){let t=typeof e=="string"?{title:e}:e;return i({__tauriModule:"Dialog",message:{cmd:"messageDialog",message:n.toString(),title:t?.title?.toString(),type:t?.type,buttonLabel:t?.okLabel?.toString()}})}async function Te(n,e){let t=typeof e=="string"?{title:e}:e;return i({__tauriModule:"Dialog",message:{cmd:"askDialog",message:n.toString(),title:t?.title?.toString(),type:t?.type,buttonLabels:[t?.okLabel?.toString()??"Yes",t?.cancelLabel?.toString()??"No"]}})}async function Ee(n,e){let t=typeof e=="string"?{title:e}:e;return i({__tauriModule:"Dialog",message:{cmd:"confirmDialog",message:n.toString(),title:t?.title?.toString(),type:t?.type,buttonLabels:[t?.okLabel?.toString()??"Ok",t?.cancelLabel?.toString()??"Cancel"]}})}var z={};c(z,{TauriEvent:()=>O,emit:()=>T,listen:()=>I,once:()=>F});async function ee(n,e){return i({__tauriModule:"Event",message:{cmd:"unlisten",event:n,eventId:e}})}async function w(n,e,t){await i({__tauriModule:"Event",message:{cmd:"emit",event:n,windowLabel:e,payload:t}})}async function _(n,e,t){return i({__tauriModule:"Event",message:{cmd:"listen",event:n,windowLabel:e,handler:m(t)}}).then(r=>async()=>ee(n,r))}async function v(n,e,t){return _(n,e,r=>{t(r),ee(n,r.id).catch(()=>{})})}var O=(d=>(d.WINDOW_RESIZED="tauri://resize",d.WINDOW_MOVED="tauri://move",d.WINDOW_CLOSE_REQUESTED="tauri://close-requested",d.WINDOW_CREATED="tauri://window-created",d.WINDOW_DESTROYED="tauri://destroyed",d.WINDOW_FOCUS="tauri://focus",d.WINDOW_BLUR="tauri://blur",d.WINDOW_SCALE_FACTOR_CHANGED="tauri://scale-change",d.WINDOW_THEME_CHANGED="tauri://theme-changed",d.WINDOW_FILE_DROP="tauri://file-drop",d.WINDOW_FILE_DROP_HOVER="tauri://file-drop-hover",d.WINDOW_FILE_DROP_CANCELLED="tauri://file-drop-cancelled",d.MENU="tauri://menu",d.CHECK_UPDATE="tauri://update",d.UPDATE_AVAILABLE="tauri://update-available",d.INSTALL_UPDATE="tauri://update-install",d.STATUS_UPDATE="tauri://update-status",d.DOWNLOAD_PROGRESS="tauri://update-download-progress",d))(O||{});async function I(n,e){return _(n,null,e)}async function F(n,e){return v(n,null,e)}async function T(n,e){return w(n,void 0,e)}var H={};c(H,{isRegistered:()=>Ae,register:()=>Ce,registerAll:()=>Me,unregister:()=>Se,unregisterAll:()=>We});async function Ce(n,e){return i({__tauriModule:"GlobalShortcut",message:{cmd:"register",shortcut:n,handler:m(e)}})}async function Me(n,e){return i({__tauriModule:"GlobalShortcut",message:{cmd:"registerAll",shortcuts:n,handler:m(e)}})}async function Ae(n){return i({__tauriModule:"GlobalShortcut",message:{cmd:"isRegistered",shortcut:n}})}async function Se(n){return i({__tauriModule:"GlobalShortcut",message:{cmd:"unregister",shortcut:n}})}async function We(){return i({__tauriModule:"GlobalShortcut",message:{cmd:"unregisterAll"}})}var q={};c(q,{Body:()=>g,Client:()=>D,Response:()=>E,ResponseType:()=>te,fetch:()=>xe,getClient:()=>ne});var te=(r=>(r[r.JSON=1]="JSON",r[r.Text=2]="Text",r[r.Binary=3]="Binary",r))(te||{}),g=class{constructor(e,t){this.type=e,this.payload=t}static form(e){let t={},r=(s,a)=>{if(a!==null){let u;typeof a=="string"?u=a:a instanceof Uint8Array||Array.isArray(a)?u=Array.from(a):a instanceof File?u={file:a.name,mime:a.type,fileName:a.name}:typeof a.file=="string"?u={file:a.file,mime:a.mime,fileName:a.fileName}:u={file:Array.from(a.file),mime:a.mime,fileName:a.fileName},t[String(s)]=u}};if(e instanceof FormData)for(let[s,a]of e)r(s,a);else for(let[s,a]of Object.entries(e))r(s,a);return new g("Form",t)}static json(e){return new g("Json",e)}static text(e){return new g("Text",e)}static bytes(e){return new g("Bytes",Array.from(e instanceof ArrayBuffer?new Uint8Array(e):e))}},E=class{constructor(e){this.url=e.url,this.status=e.status,this.ok=this.status>=200&&this.status<300,this.headers=e.headers,this.rawHeaders=e.rawHeaders,this.data=e.data}},D=class{constructor(e){this.id=e}async drop(){return i({__tauriModule:"Http",message:{cmd:"dropClient",client:this.id}})}async request(e){let t=!e.responseType||e.responseType===1;return t&&(e.responseType=2),i({__tauriModule:"Http",message:{cmd:"httpRequest",client:this.id,options:e}}).then(r=>{let s=new E(r);if(t){try{s.data=JSON.parse(s.data)}catch(a){if(s.ok&&s.data==="")s.data={};else if(s.ok)throw Error(`Failed to parse response \`${s.data}\` as JSON: ${a};
+              try setting the \`responseType\` option to \`ResponseType.Text\` or \`ResponseType.Binary\` if the API does not return a JSON response.`)}return s}return s})}async get(e,t){return this.request({method:"GET",url:e,...t})}async post(e,t,r){return this.request({method:"POST",url:e,body:t,...r})}async put(e,t,r){return this.request({method:"PUT",url:e,body:t,...r})}async patch(e,t){return this.request({method:"PATCH",url:e,...t})}async delete(e,t){return this.request({method:"DELETE",url:e,...t})}};async function ne(n){return i({__tauriModule:"Http",message:{cmd:"createClient",options:n}}).then(e=>new D(e))}var V=null;async function xe(n,e){return V===null&&(V=await ne()),V.request({url:n,method:e?.method??"GET",...e})}var G={};c(G,{isPermissionGranted:()=>Le,requestPermission:()=>ke,sendNotification:()=>Re});async function Le(){return window.Notification.permission!=="default"?Promise.resolve(window.Notification.permission==="granted"):i({__tauriModule:"Notification",message:{cmd:"isNotificationPermissionGranted"}})}async function ke(){return window.Notification.requestPermission()}function Re(n){typeof n=="string"?new window.Notification(n):new window.Notification(n.title,n)}var j={};c(j,{BaseDirectory:()=>ie,appCacheDir:()=>Fe,appConfigDir:()=>Ne,appDataDir:()=>Ue,appLocalDataDir:()=>Ie,appLogDir:()=>rt,audioDir:()=>ze,basename:()=>mt,cacheDir:()=>He,configDir:()=>Ve,dataDir:()=>qe,delimiter:()=>at,desktopDir:()=>Ge,dirname:()=>dt,documentDir:()=>je,downloadDir:()=>$e,executableDir:()=>Je,extname:()=>ct,fontDir:()=>Ke,homeDir:()=>Qe,isAbsolute:()=>pt,join:()=>ut,localDataDir:()=>Ye,normalize:()=>lt,pictureDir:()=>Ze,publicDir:()=>Xe,resolve:()=>ot,resolveResource:()=>et,resourceDir:()=>Be,runtimeDir:()=>tt,sep:()=>st,templateDir:()=>nt,videoDir:()=>it});function P(){return navigator.appVersion.includes("Win")}var ie=(o=>(o[o.Audio=1]="Audio",o[o.Cache=2]="Cache",o[o.Config=3]="Config",o[o.Data=4]="Data",o[o.LocalData=5]="LocalData",o[o.Document=6]="Document",o[o.Download=7]="Download",o[o.Picture=8]="Picture",o[o.Public=9]="Public",o[o.Video=10]="Video",o[o.Resource=11]="Resource",o[o.Temp=12]="Temp",o[o.AppConfig=13]="AppConfig",o[o.AppData=14]="AppData",o[o.AppLocalData=15]="AppLocalData",o[o.AppCache=16]="AppCache",o[o.AppLog=17]="AppLog",o[o.Desktop=18]="Desktop",o[o.Executable=19]="Executable",o[o.Font=20]="Font",o[o.Home=21]="Home",o[o.Runtime=22]="Runtime",o[o.Template=23]="Template",o))(ie||{});async function Ne(){return l("plugin:path|resolve_directory",{directory:13})}async function Ue(){return l("plugin:path|resolve_directory",{directory:14})}async function Ie(){return l("plugin:path|resolve_directory",{directory:15})}async function Fe(){return l("plugin:path|resolve_directory",{directory:16})}async function ze(){return l("plugin:path|resolve_directory",{directory:1})}async function He(){return l("plugin:path|resolve_directory",{directory:2})}async function Ve(){return l("plugin:path|resolve_directory",{directory:3})}async function qe(){return l("plugin:path|resolve_directory",{directory:4})}async function Ge(){return l("plugin:path|resolve_directory",{directory:18})}async function je(){return l("plugin:path|resolve_directory",{directory:6})}async function $e(){return l("plugin:path|resolve_directory",{directory:7})}async function Je(){return l("plugin:path|resolve_directory",{directory:19})}async function Ke(){return l("plugin:path|resolve_directory",{directory:20})}async function Qe(){return l("plugin:path|resolve_directory",{directory:21})}async function Ye(){return l("plugin:path|resolve_directory",{directory:5})}async function Ze(){return l("plugin:path|resolve_directory",{directory:8})}async function Xe(){return l("plugin:path|resolve_directory",{directory:9})}async function Be(){return l("plugin:path|resolve_directory",{directory:11})}async function et(n){return l("plugin:path|resolve_directory",{directory:11,path:n})}async function tt(){return l("plugin:path|resolve_directory",{directory:22})}async function nt(){return l("plugin:path|resolve_directory",{directory:23})}async function it(){return l("plugin:path|resolve_directory",{directory:10})}async function rt(){return l("plugin:path|resolve_directory",{directory:17})}var st=P()?"\\":"/",at=P()?";":":";async function ot(...n){return l("plugin:path|resolve",{paths:n})}async function lt(n){return l("plugin:path|normalize",{path:n})}async function ut(...n){return l("plugin:path|join",{paths:n})}async function dt(n){return l("plugin:path|dirname",{path:n})}async function ct(n){return l("plugin:path|extname",{path:n})}async function mt(n,e){return l("plugin:path|basename",{path:n,ext:e})}async function pt(n){return l("plugin:path|isAbsolute",{path:n})}var $={};c($,{exit:()=>gt,relaunch:()=>yt});async function gt(n=0){return i({__tauriModule:"Process",message:{cmd:"exit",exitCode:n}})}async function yt(){return i({__tauriModule:"Process",message:{cmd:"relaunch"}})}var J={};c(J,{Child:()=>C,Command:()=>h,EventEmitter:()=>y,open:()=>ft});async function ht(n,e,t=[],r){return typeof t=="object"&&Object.freeze(t),i({__tauriModule:"Shell",message:{cmd:"execute",program:e,args:t,options:r,onEventFn:m(n)}})}var y=class{constructor(){this.eventListeners=Object.create(null)}addListener(e,t){return this.on(e,t)}removeListener(e,t){return this.off(e,t)}on(e,t){return e in this.eventListeners?this.eventListeners[e].push(t):this.eventListeners[e]=[t],this}once(e,t){let r=s=>{this.removeListener(e,r),t(s)};return this.addListener(e,r)}off(e,t){return e in this.eventListeners&&(this.eventListeners[e]=this.eventListeners[e].filter(r=>r!==t)),this}removeAllListeners(e){return e?delete this.eventListeners[e]:this.eventListeners=Object.create(null),this}emit(e,t){if(e in this.eventListeners){let r=this.eventListeners[e];for(let s of r)s(t);return!0}return!1}listenerCount(e){return e in this.eventListeners?this.eventListeners[e].length:0}prependListener(e,t){return e in this.eventListeners?this.eventListeners[e].unshift(t):this.eventListeners[e]=[t],this}prependOnceListener(e,t){let r=s=>{this.removeListener(e,r),t(s)};return this.prependListener(e,r)}},C=class{constructor(e){this.pid=e}async write(e){return i({__tauriModule:"Shell",message:{cmd:"stdinWrite",pid:this.pid,buffer:typeof e=="string"?e:Array.from(e)}})}async kill(){return i({__tauriModule:"Shell",message:{cmd:"killChild",pid:this.pid}})}},h=class extends y{constructor(t,r=[],s){super();this.stdout=new y;this.stderr=new y;this.program=t,this.args=typeof r=="string"?[r]:r,this.options=s??{}}static create(t,r=[],s){return new h(t,r,s)}static sidecar(t,r=[],s){let a=new h(t,r,s);return a.options.sidecar=!0,a}async spawn(){return ht(t=>{switch(t.event){case"Error":this.emit("error",t.payload);break;case"Terminated":this.emit("close",t.payload);break;case"Stdout":this.stdout.emit("data",t.payload);break;case"Stderr":this.stderr.emit("data",t.payload);break}},this.program,this.args,this.options).then(t=>new C(t))}async execute(){return new Promise((t,r)=>{this.on("error",r);let s=[],a=[];this.stdout.on("data",u=>{s.push(u)}),this.stderr.on("data",u=>{a.push(u)}),this.on("close",u=>{t({code:u.code,signal:u.signal,stdout:this.collectOutput(s),stderr:this.collectOutput(a)})}),this.spawn().catch(r)})}collectOutput(t){return this.options.encoding==="raw"?t.reduce((r,s)=>new Uint8Array([...r,...s,10]),new Uint8Array):t.join(`
+`)}};async function ft(n,e){return i({__tauriModule:"Shell",message:{cmd:"open",path:n,with:e}})}var Q={};c(Q,{checkUpdate:()=>_t,installUpdate:()=>bt,onUpdaterEvent:()=>K});async function K(n){return I("tauri://update-status",e=>{n(e?.payload)})}async function bt(){let n;function e(){n&&n(),n=void 0}return new Promise((t,r)=>{function s(a){if(a.error){e(),r(a.error);return}a.status==="DONE"&&(e(),t())}K(s).then(a=>{n=a}).catch(a=>{throw e(),a}),T("tauri://update-install").catch(a=>{throw e(),a})})}async function _t(){let n;function e(){n&&n(),n=void 0}return new Promise((t,r)=>{function s(u){e(),t({manifest:u,shouldUpdate:!0})}function a(u){if(u.error){e(),r(u.error);return}u.status==="UPTODATE"&&(e(),t({shouldUpdate:!1}))}F("tauri://update-available",u=>{s(u?.payload)}).catch(u=>{throw e(),u}),K(a).then(u=>{n=u}).catch(u=>{throw e(),u}),T("tauri://update").catch(u=>{throw e(),u})})}var X={};c(X,{CloseRequestedEvent:()=>x,LogicalPosition:()=>A,LogicalSize:()=>M,PhysicalPosition:()=>b,PhysicalSize:()=>f,UserAttentionType:()=>se,WebviewWindow:()=>p,WebviewWindowHandle:()=>S,WindowManager:()=>W,appWindow:()=>Y,availableMonitors:()=>Ot,currentMonitor:()=>wt,getAll:()=>ae,getCurrent:()=>Pt,primaryMonitor:()=>vt});var M=class{constructor(e,t){this.type="Logical";this.width=e,this.height=t}},f=class{constructor(e,t){this.type="Physical";this.width=e,this.height=t}toLogical(e){return new M(this.width/e,this.height/e)}},A=class{constructor(e,t){this.type="Logical";this.x=e,this.y=t}},b=class{constructor(e,t){this.type="Physical";this.x=e,this.y=t}toLogical(e){return new A(this.x/e,this.y/e)}},se=(t=>(t[t.Critical=1]="Critical",t[t.Informational=2]="Informational",t))(se||{});function Pt(){return new p(window.__TAURI_METADATA__.__currentWindow.label,{skip:!0})}function ae(){return window.__TAURI_METADATA__.__windows.map(n=>new p(n.label,{skip:!0}))}var re=["tauri://created","tauri://error"],S=class{constructor(e){this.label=e,this.listeners=Object.create(null)}async listen(e,t){return this._handleTauriEvent(e,t)?Promise.resolve(()=>{let r=this.listeners[e];r.splice(r.indexOf(t),1)}):_(e,this.label,t)}async once(e,t){return this._handleTauriEvent(e,t)?Promise.resolve(()=>{let r=this.listeners[e];r.splice(r.indexOf(t),1)}):v(e,this.label,t)}async emit(e,t){if(re.includes(e)){for(let r of this.listeners[e]||[])r({event:e,id:-1,windowLabel:this.label,payload:t});return Promise.resolve()}return w(e,this.label,t)}_handleTauriEvent(e,t){return re.includes(e)?(e in this.listeners?this.listeners[e].push(t):this.listeners[e]=[t],!0):!1}},W=class extends S{async scaleFactor(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"scaleFactor"}}}})}async innerPosition(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"innerPosition"}}}}).then(({x:e,y:t})=>new b(e,t))}async outerPosition(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"outerPosition"}}}}).then(({x:e,y:t})=>new b(e,t))}async innerSize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"innerSize"}}}}).then(({width:e,height:t})=>new f(e,t))}async outerSize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"outerSize"}}}}).then(({width:e,height:t})=>new f(e,t))}async isFullscreen(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isFullscreen"}}}})}async isMinimized(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isMinimized"}}}})}async isMaximized(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isMaximized"}}}})}async isDecorated(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isDecorated"}}}})}async isResizable(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isResizable"}}}})}async isVisible(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"isVisible"}}}})}async title(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"title"}}}})}async theme(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"theme"}}}})}async center(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"center"}}}})}async requestUserAttention(e){let t=null;return e&&(e===1?t={type:"Critical"}:t={type:"Informational"}),i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"requestUserAttention",payload:t}}}})}async setResizable(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setResizable",payload:e}}}})}async setTitle(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setTitle",payload:e}}}})}async maximize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"maximize"}}}})}async unmaximize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"unmaximize"}}}})}async toggleMaximize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"toggleMaximize"}}}})}async minimize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"minimize"}}}})}async unminimize(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"unminimize"}}}})}async show(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"show"}}}})}async hide(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"hide"}}}})}async close(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"close"}}}})}async setDecorations(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setDecorations",payload:e}}}})}async setShadow(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setShadow",payload:e}}}})}async setAlwaysOnTop(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setAlwaysOnTop",payload:e}}}})}async setContentProtected(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setContentProtected",payload:e}}}})}async setSize(e){if(!e||e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `size` argument must be either a LogicalSize or a PhysicalSize instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setSize",payload:{type:e.type,data:{width:e.width,height:e.height}}}}}})}async setMinSize(e){if(e&&e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `size` argument must be either a LogicalSize or a PhysicalSize instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setMinSize",payload:e?{type:e.type,data:{width:e.width,height:e.height}}:null}}}})}async setMaxSize(e){if(e&&e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `size` argument must be either a LogicalSize or a PhysicalSize instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setMaxSize",payload:e?{type:e.type,data:{width:e.width,height:e.height}}:null}}}})}async setPosition(e){if(!e||e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `position` argument must be either a LogicalPosition or a PhysicalPosition instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setPosition",payload:{type:e.type,data:{x:e.x,y:e.y}}}}}})}async setFullscreen(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setFullscreen",payload:e}}}})}async setFocus(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setFocus"}}}})}async setIcon(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setIcon",payload:{icon:typeof e=="string"?e:Array.from(e)}}}}})}async setSkipTaskbar(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setSkipTaskbar",payload:e}}}})}async setCursorGrab(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorGrab",payload:e}}}})}async setCursorVisible(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorVisible",payload:e}}}})}async setCursorIcon(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorIcon",payload:e}}}})}async setCursorPosition(e){if(!e||e.type!=="Logical"&&e.type!=="Physical")throw new Error("the `position` argument must be either a LogicalPosition or a PhysicalPosition instance");return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setCursorPosition",payload:{type:e.type,data:{x:e.x,y:e.y}}}}}})}async setIgnoreCursorEvents(e){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"setIgnoreCursorEvents",payload:e}}}})}async startDragging(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{label:this.label,cmd:{type:"startDragging"}}}})}async onResized(e){return this.listen("tauri://resize",e)}async onMoved(e){return this.listen("tauri://move",e)}async onCloseRequested(e){return this.listen("tauri://close-requested",t=>{let r=new x(t);Promise.resolve(e(r)).then(()=>{if(!r.isPreventDefault())return this.close()})})}async onFocusChanged(e){let t=await this.listen("tauri://focus",s=>{e({...s,payload:!0})}),r=await this.listen("tauri://blur",s=>{e({...s,payload:!1})});return()=>{t(),r()}}async onScaleChanged(e){return this.listen("tauri://scale-change",e)}async onMenuClicked(e){return this.listen("tauri://menu",e)}async onFileDropEvent(e){let t=await this.listen("tauri://file-drop",a=>{e({...a,payload:{type:"drop",paths:a.payload}})}),r=await this.listen("tauri://file-drop-hover",a=>{e({...a,payload:{type:"hover",paths:a.payload}})}),s=await this.listen("tauri://file-drop-cancelled",a=>{e({...a,payload:{type:"cancel"}})});return()=>{t(),r(),s()}}async onThemeChanged(e){return this.listen("tauri://theme-changed",e)}},x=class{constructor(e){this._preventDefault=!1;this.event=e.event,this.windowLabel=e.windowLabel,this.id=e.id}preventDefault(){this._preventDefault=!0}isPreventDefault(){return this._preventDefault}},p=class extends W{constructor(e,t={}){super(e),t?.skip||i({__tauriModule:"Window",message:{cmd:"createWebview",data:{options:{label:e,...t}}}}).then(async()=>this.emit("tauri://created")).catch(async r=>this.emit("tauri://error",r))}static getByLabel(e){return ae().some(t=>t.label===e)?new p(e,{skip:!0}):null}},Y;"__TAURI_METADATA__"in window?Y=new p(window.__TAURI_METADATA__.__currentWindow.label,{skip:!0}):(console.warn(`Could not find "window.__TAURI_METADATA__". The "appWindow" value will reference the "main" window label.
+Note that this is not an issue if running this frontend on a browser instead of a Tauri window.`),Y=new p("main",{skip:!0}));function Z(n){return n===null?null:{name:n.name,scaleFactor:n.scaleFactor,position:new b(n.position.x,n.position.y),size:new f(n.size.width,n.size.height)}}async function wt(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{cmd:{type:"currentMonitor"}}}}).then(Z)}async function vt(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{cmd:{type:"primaryMonitor"}}}}).then(Z)}async function Ot(){return i({__tauriModule:"Window",message:{cmd:"manage",data:{cmd:{type:"availableMonitors"}}}}).then(n=>n.map(Z))}var B={};c(B,{EOL:()=>Tt,arch:()=>Mt,platform:()=>Et,tempdir:()=>At,type:()=>Ct,version:()=>Dt});var Tt=P()?`\r
 `:`
-`;async function Ut(){return i({__tauriModule:"Os",message:{cmd:"platform"}})}async function Nt(){return i({__tauriModule:"Os",message:{cmd:"version"}})}async function It(){return i({__tauriModule:"Os",message:{cmd:"osType"}})}async function zt(){return i({__tauriModule:"Os",message:{cmd:"arch"}})}async function Ht(){return i({__tauriModule:"Os",message:{cmd:"tempdir"}})}var Vt=l;return me(jt);})();
+`;async function Et(){return i({__tauriModule:"Os",message:{cmd:"platform"}})}async function Dt(){return i({__tauriModule:"Os",message:{cmd:"version"}})}async function Ct(){return i({__tauriModule:"Os",message:{cmd:"osType"}})}async function Mt(){return i({__tauriModule:"Os",message:{cmd:"arch"}})}async function At(){return i({__tauriModule:"Os",message:{cmd:"tempdir"}})}var St=l;return ce(Wt);})();
 window.__TAURI__ = __TAURI_IIFE__
